fix: address xeno review feedback
